+++ src/test/test-utils/test-data-loader.ts
@@ -1,7 +1,0 @@
-import * as fs from "fs";
-
-export class TestDataLoader {
-    static load(fileName: string): string {
-        return fs.readFileSync(`${__dirname}/../../../src/test/test-data/${fileName}`, "utf-8");
-    }
-}