+++ src/xpath/index.ts
@@ -1,2 +1,0 @@
-export * from "./xpath-builder";
-export * from "./xpath-evaluator";